+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased] - [Work in Progress][v4]
 
   - Updated to Truffle 0.25
+  - Fixed trace replay ([PR #127](https://github.com/smarr/SOMns/pull/127))
 
 ## [0.3.0] - [2017-04-06 &lt;Programming&gt;'17 Demo][v3]